treewide: plug memleak: case 2.2.2: for_each_child_of_node() (v6.12 prep)
Case 2.2.2 from below bug:

    If returning or breaking early from within
    for_each_child_of_node(), a reference to the current device_node
    remains and needs to be dropped, either right away, or later if
    used past that loop to avoid a memleak.

In this case, the np is not needed past the loop and code can be
simplified by switching to for_each_child_of_node_scoped() which avoids
the need for manual cleanup in early exits from the loop.

Note that the error is in dsim_of_get_pll_diags() only, we also update
dsim_of_get_clock_mode() as well to have aligned code in here, though.

This commit plugs memleaks in error paths.

Test: TH
Bug: 370679843
Change-Id: I1f558eb068c277b50fe8b10623cfb3f679d91f1e
Signed-off-by: André Draszik <draszik@google.com>