treewide: plug memleak: case 2.2.1: for_each_available_child_of_node() (v6.12 prep)
Case 2.2.1 from below bug:

    If returning or breaking early from within
    for_each_available_child_of_node(), a reference to the current
    device_node remains and needs to be dropped, either right away, or
    later if used past that loop to avoid a memleak.

In this case, an early exit from the loop is missing the call to
of_node_put() and code can be simplified by switching to
for_each_available_child_of_node_scoped() which avoids the need for
manual cleanup in early exits from the loop.

This commit plugs a memleak in an error path.

Test: TH
Bug: 370679843
Change-Id: Ia103c8cf477e2f690e7f74ede85d7bcb9fc33532
Signed-off-by: André Draszik <draszik@google.com>